taskflow-api/requirements.txt
asgiref==3.8.1
attrs==25.4.0
Django==5.1.4
django-cors-headers==4.9.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
setuptools==80.3.1
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.2.0